refactor: rewrite voice state update to TypeScript
